configs: socfpga: Add routine to update HPS EMIF I/O
For Agilex, the HPS EMIF I/O peripheral rbf is stored
in SDM flash. We need a method to update the flash with
a peripheral configuration stored on the SD card.

To do this, we encode a peripheral version ID in the SDM
flash at offset 0x300000. Read the value at this offset
from the file on the SD card and from SDM flash, and if they
do not match, then rewrite SDM flash and reset the HPS.

Signed-off-by: Tom Mealey <tmealey@mathworks.com>